refactor: move endpoints utils file higher
